fix: correct rename of files
