+++ packages/client/src/hooks/useAuthCheck.ts
